misc/HtmlTestSuite/results/test-1301.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Placing selection inside empty elements</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div contenteditable="" id="host">
  <p><strong id="strong"/></p>
</div>
<script>
test( () =&gt; {
  const range = document.createRange();

  range.setStart( strong, 0 );
  range.collapse()
  getSelection().removeAllRanges();
  getSelection().addRange( range );

  const selectedRange = getSelection().getRangeAt( 0 );

  assert_equals( selectedRange.startContainer, strong );
  assert_equals( selectedRange.startOffset, 0 );
  assert_equals( selectedRange.endContainer, strong );
  assert_equals( selectedRange.endOffset, 0 );
}, 'Selection can be placed inside the empty element' );
</script>
</body></html>
